fix tuple constructors parameter type

+++ inc/minirt.h
@@ -6,7 +6,7 @@
 /*   By: maolivei <maolivei@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/15 10:24:24 by ralves-b          #+#    #+#             */
-/*   Updated: 2022/10/17 16:51:40 by maolivei         ###   ########.fr       */
+/*   Updated: 2022/10/17 21:32:36 by maolivei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -32,9 +32,9 @@ typedef struct s_tuple
 typedef t_tuple	t_vector; /* tuple with 'w' value equal to 0 */
 typedef t_tuple	t_point; /* tuple with 'w' value equal to 1 */
 
-t_tuple		*create_tuple(int x, int y, int z, int w);
-t_point		*create_point(int x, int y, int z);
-t_vector	*create_vector(int x, int y, int z);
+t_tuple		*create_tuple(double x, double y, double z, double w);
+t_point		*create_point(double x, double y, double z);
+t_vector	*create_vector(double x, double y, double z);
 
 t_bool		is_equal(double a, double b);
 t_bool		is_equal_tuple(t_tuple *a, t_tuple *b);
